+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -16,6 +16,7 @@
 #include "Rule10p6Check.h"
 #include "Rule10p7Check.h"
 #include "Rule10p8Check.h"
+#include "Rule11p1Check.h"
 #include "Rule11p5Check.h"
 #include "Rule11p8Check.h"
 #include "Rule11p9Check.h"
@@ -81,6 +82,8 @@ public:
         "misrac-rule-10p7");
     CheckFactories.registerCheck<Rule10p8Check>(
         "misrac-rule-10p8");
+    CheckFactories.registerCheck<Rule11p1Check>(
+        "misrac-rule-11p1");
     CheckFactories.registerCheck<Rule11p5Check>(
         "misrac-rule-11p5");
     CheckFactories.registerCheck<Rule11p8Check>(